Move popgetter cache to lib with cache feature flag
Moves the constructor that uses cached metadata from cli to lib crates
to enable its use from the Python bindings crate.
